requirements.txt
langchain==0.2.3
langchain-community==0.2.4
langchain-core==0.2.5
langchain-text-splitters==0.2.1
langsmith==0.1.75
numpy==1.26.4
openai==1.33.0
pandas==2.2.1
python-dotenv==1.0.1
youtube-transcript-api==0.6.2
pypdf==4.2.0
